Move demo files into stories directory
