+++ packages/components/organisms/f-checkout/src/store/tests/checkout.module.test.js
@@ -1,12 +1,12 @@
 import axios from 'axios';
 import CheckoutModule from '../checkout.module';
-import checkoutDelivery from '../../demo/checkout-delivery.json';
-import basketDelivery from '../../demo/get-basket-delivery.json';
-import basketDeliveryAgeRestricted from '../../demo/get-basket-delivery-age-restriction.json';
-import checkoutAvailableFulfilment from '../../demo/checkout-available-fulfilment.json';
-import customerAddresses from '../../demo/get-address.json';
-import geoLocationDetails from '../../demo/get-geo-location.json';
-import customer from '../../demo/get-customer.json';
+import checkoutDelivery from '../../../stories/demo/checkout-delivery.json';
+import basketDelivery from '../../../stories/demo/get-basket-delivery.json';
+import basketDeliveryAgeRestricted from '../../../stories/demo/get-basket-delivery-age-restriction.json';
+import checkoutAvailableFulfilment from '../../../stories/demo/checkout-available-fulfilment.json';
+import customerAddresses from '../../../stories/demo/get-address.json';
+import geoLocationDetails from '../../../stories/demo/get-geo-location.json';
+import customer from '../../../stories/demo/get-customer.json';
 import storageMock from '../../../test-utils/local-storage/local-storage-mock';
 import addressService from '../../services/addressService';
 import basketApi from '../../services/basketApi';
